+++ src/main/java/ru/yandex/practicum/filmorate/repository/MpaRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface MpaRepository {
     List<Mpa> findAll();
 
-    Optional<Mpa> findById(Long id);
+    Optional<Mpa> findById(int id);
 }
