+++ src/main/java/ru/yandex/practicum/filmorate/service/GenreService.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 public interface GenreService {
     List<Genre> findAll();
 
-    Optional<Genre> findById(Long id);
+    Optional<Genre> findById(int id);
 }
